1. Refactor existing Views to use Material Design components 1. Added navigation bar Material Design
