+++ app/src/main/java/com/vazzab/vfplay/BookmarksFragment.kt
@@ -1,0 +1,14 @@
+package com.vazzab.vfplay
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+class BookmarksFragment : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_bookmarks, container, false)
+    }
+}
