+++ app/src/main/java/com/vazzab/vfplay/MainActivity.kt
@@ -1,27 +1,26 @@
 package com.vazzab.vfplay
 
 import android.os.Bundle
-import android.widget.Button
-import android.widget.Toast
-import androidx.activity.ComponentActivity
+import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.NavigationUI
+import com.google.android.material.bottomnavigation.BottomNavigationView
 
-class MainActivity : ComponentActivity() {
+
+class MainActivity : AppCompatActivity() {
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val buttons = listOf(
-            R.id.button1 to R.string.navigation_button_main,
-            R.id.button2 to R.string.navigation_button_category,
-            R.id.button3 to R.string.navigation_button_bookmark,
-            R.id.button4 to R.string.navigation_button_new,
-            R.id.button5 to R.string.navigation_button_top
-        )
-
-        buttons.forEach { (buttonId, stringId) ->
-            findViewById<Button>(buttonId).setOnClickListener {
-                Toast.makeText(this, getString(stringId), Toast.LENGTH_SHORT).show()
-            }
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment_container) as? NavHostFragment
+        if (navHostFragment != null) {
+            val navController = navHostFragment.navController
+            val bottomNavView = findViewById<BottomNavigationView>(R.id.bottom_nav_view)
+            NavigationUI.setupWithNavController(bottomNavView, navController)
+        } else {
+            Log.e("MainActivity", "NavHostFragment is null. Check activity_main.xml")
         }
 
     }
