+++ app/src/main/java/com/vazzab/vfplay/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
 package com.vazzab.vfplay.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
