+++ app/src/main/res/drawable/ic_categories.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:tint="#000000" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
+      
+    <path android:fillColor="@android:color/white" android:pathData="M12,2l-5.5,9h11z"/>
+      
+    <path android:fillColor="@android:color/white" android:pathData="M17.5,17.5m-4.5,0a4.5,4.5 0,1 1,9 0a4.5,4.5 0,1 1,-9 0"/>
+      
+    <path android:fillColor="@android:color/white" android:pathData="M3,13.5h8v8H3z"/>
+    
+</vector>
