+++ app/src/main/res/layout/activity_main.xml
@@ -1,27 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/teal_200">
-
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/top_bar_card"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_margin="5dp"
-
-        android:background="@color/main_screen"
-        app:cardCornerRadius="10dp"
-        app:cardElevation="15dp"
-
-        app:layout_constraintBottom_toTopOf="@id/title_container"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-    </androidx.cardview.widget.CardView>
+    android:layout_height="match_parent">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/title_container"
@@ -29,12 +11,7 @@
         android:layout_height="100dp"
         android:layout_margin="5dp"
         app:cardCornerRadius="10dp"
-        app:cardElevation="15dp"
-
-        app:layout_constraintBottom_toTopOf="@id/main_screen"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/top_bar_card">
+        app:cardElevation="15dp">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -85,116 +62,20 @@
         </LinearLayout>
     </androidx.cardview.widget.CardView>
 
-    <LinearLayout
-        android:id="@+id/main_screen"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:background="@color/main_screen"
-        android:orientation="vertical"
-        android:layout_margin="5dp"
-
-        app:layout_constraintBottom_toTopOf="@id/bottom_bar_card"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/title_container"
-        app:layout_constraintVertical_bias="0.0">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="5dp"
-            android:linksClickable="true"
-            android:autoLink="all"
-            android:text="@string/main_screen_text"
-            android:textSize="20sp"
-            android:textColor="@color/black"
-            android:textStyle="bold"
-
-
-            />
-
-
-    </LinearLayout>
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/bottom_bar_card"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/nav_host_fragment_container"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:layout_margin="5dp"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/nav_graph" />
 
-        app:cardCornerRadius="10dp"
-        app:cardElevation="15dp"
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/bottom_nav_view"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom"
+        app:menu="@menu/bottom_nav_menu" />
 
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-
-        <LinearLayout
-            style="?android:attr/buttonBarStyle"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-
-            <Button
-                android:id="@+id/button1"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?android:attr/selectableItemBackgroundBorderless"
-
-                android:text="@string/navigation_button_main"
-                android:textSize="@dimen/main_text_size" />
-
-
-            <Button
-                android:id="@+id/button2"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?android:attr/selectableItemBackgroundBorderless"
-
-                android:text="@string/navigation_button_category"
-                android:textSize="@dimen/main_text_size" />
-
-            <Button
-                android:id="@+id/button3"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?android:attr/selectableItemBackgroundBorderless"
-
-
-                android:text="@string/navigation_button_bookmark"
-                android:textSize="@dimen/main_text_size" />
-
-            <Button
-                android:id="@+id/button4"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?android:attr/selectableItemBackgroundBorderless"
-
-                android:text="@string/navigation_button_new"
-                android:textSize="@dimen/main_text_size" />
-
-            <Button
-                android:id="@+id/button5"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?android:attr/selectableItemBackgroundBorderless"
-
-                android:text="@string/navigation_button_top"
-                android:textSize="@dimen/main_text_size" />
-        </LinearLayout>
-
-
-    </androidx.cardview.widget.CardView>
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
