--- app/src/main/res/values/dimen.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <dimen name="main_text_size">10sp</dimen>
-</resources>
