+++ app/src/main/res/values/themes.xml
@@ -1,5 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <resources>
-
-    <style name="Theme.VFplay" parent="android:Theme.Material.Light.NoActionBar" />
+    <style name="Theme.VFplay" parent="Theme.AppCompat.Light.NoActionBar" />
 </resources>
